Extract player adapters
